+++ ReleaseNotes.md
@@ -1,4 +1,6 @@
 # v4.NEXT
+### Features 
+- cta/CTA#16 - Add option for a user config file
 
 ## Summary
 ### Features